+++ docs/html/search/classes_0.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['clockinterface_58',['ClockInterface',['../classace__segment_1_1ClockInterface.html',1,'ace_segment']]]
+  ['clockinterface_60',['ClockInterface',['../classace__segment_1_1ClockInterface.html',1,'ace_segment']]]
 ];