+++ docs/html/search/classes_1.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['directfast4module_59',['DirectFast4Module',['../classace__segment_1_1DirectFast4Module.html',1,'ace_segment']]],
-  ['directmodule_60',['DirectModule',['../classace__segment_1_1DirectModule.html',1,'ace_segment']]]
+  ['directfast4module_61',['DirectFast4Module',['../classace__segment_1_1DirectFast4Module.html',1,'ace_segment']]],
+  ['directmodule_62',['DirectModule',['../classace__segment_1_1DirectModule.html',1,'ace_segment']]]
 ];