+++ docs/html/search/classes_3.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['hc595module_62',['Hc595Module',['../classace__segment_1_1Hc595Module.html',1,'ace_segment']]],
-  ['ht16k33module_63',['Ht16k33Module',['../classace__segment_1_1Ht16k33Module.html',1,'ace_segment']]],
-  ['hybridmodule_64',['HybridModule',['../classace__segment_1_1HybridModule.html',1,'ace_segment']]]
+  ['hc595module_64',['Hc595Module',['../classace__segment_1_1Hc595Module.html',1,'ace_segment']]],
+  ['ht16k33module_65',['Ht16k33Module',['../classace__segment_1_1Ht16k33Module.html',1,'ace_segment']]],
+  ['hybridmodule_66',['HybridModule',['../classace__segment_1_1HybridModule.html',1,'ace_segment']]]
 ];