+++ docs/html/search/functions_2.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['digitalwrite_85',['digitalWrite',['../classace__segment_1_1GpioInterface.html#ae87dcb5fa1c1245fc5747abfeebac7b4',1,'ace_segment::GpioInterface']]],
-  ['disablegroup_86',['disableGroup',['../classace__segment_1_1LedMatrixBase.html#a392418ec3a4e763920e21acc763e8848',1,'ace_segment::LedMatrixBase::disableGroup()'],['../classace__segment_1_1LedMatrixDualHc595.html#a0a08cfdd1de0cbdb20b9e66b22627c86',1,'ace_segment::LedMatrixDualHc595::disableGroup()']]],
-  ['draw_87',['draw',['../classace__segment_1_1LedMatrixBase.html#ad9d2bc50b50ab4b2c16c1ef53b3397d2',1,'ace_segment::LedMatrixBase::draw()'],['../classace__segment_1_1LedMatrixDualHc595.html#ac191206242ca55ebb31652de7993277e',1,'ace_segment::LedMatrixDualHc595::draw()']]]
+  ['digitalwrite_88',['digitalWrite',['../classace__segment_1_1GpioInterface.html#ae87dcb5fa1c1245fc5747abfeebac7b4',1,'ace_segment::GpioInterface']]],
+  ['disablegroup_89',['disableGroup',['../classace__segment_1_1LedMatrixBase.html#a392418ec3a4e763920e21acc763e8848',1,'ace_segment::LedMatrixBase::disableGroup()'],['../classace__segment_1_1LedMatrixDualHc595.html#a0a08cfdd1de0cbdb20b9e66b22627c86',1,'ace_segment::LedMatrixDualHc595::disableGroup()']]],
+  ['draw_90',['draw',['../classace__segment_1_1LedMatrixBase.html#ad9d2bc50b50ab4b2c16c1ef53b3397d2',1,'ace_segment::LedMatrixBase::draw()'],['../classace__segment_1_1LedMatrixDualHc595.html#ac191206242ca55ebb31652de7993277e',1,'ace_segment::LedMatrixDualHc595::draw()']]]
 ];