+++ docs/html/search/functions_4.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['flush_91',['flush',['../classace__segment_1_1Ht16k33Module.html#ad9225465a7ce3fa621eb8505edb77f6a',1,'ace_segment::Ht16k33Module::flush()'],['../classace__segment_1_1Max7219Module.html#af8b588c48e73e6883948eb8159982a1c',1,'ace_segment::Max7219Module::flush()'],['../classace__segment_1_1Tm1637Module.html#af3a100b0d07125af76aaf8ea10bc597c',1,'ace_segment::Tm1637Module::flush()'],['../classace__segment_1_1Tm1638Module.html#a76c67711f28cc6b5eba38681b62c4d4d',1,'ace_segment::Tm1638Module::flush()']]],
-  ['flushincremental_92',['flushIncremental',['../classace__segment_1_1Tm1637Module.html#a1bf9dcdafb5b1a08a18da697d29e5235',1,'ace_segment::Tm1637Module']]]
+  ['flush_94',['flush',['../classace__segment_1_1Ht16k33Module.html#ad9225465a7ce3fa621eb8505edb77f6a',1,'ace_segment::Ht16k33Module::flush()'],['../classace__segment_1_1Max7219Module.html#af8b588c48e73e6883948eb8159982a1c',1,'ace_segment::Max7219Module::flush()'],['../classace__segment_1_1Tm1637Module.html#af3a100b0d07125af76aaf8ea10bc597c',1,'ace_segment::Tm1637Module::flush()'],['../classace__segment_1_1Tm1638AnodeModule.html#a7b94991afe9a3146bd7d418ff3c516b3',1,'ace_segment::Tm1638AnodeModule::flush()'],['../classace__segment_1_1Tm1638Module.html#a76c67711f28cc6b5eba38681b62c4d4d',1,'ace_segment::Tm1638Module::flush()']]],
+  ['flushincremental_95',['flushIncremental',['../classace__segment_1_1Tm1637Module.html#a1bf9dcdafb5b1a08a18da697d29e5235',1,'ace_segment::Tm1637Module']]]
 ];