+++ docs/html/search/functions_a.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['pinmode_114',['pinMode',['../classace__segment_1_1GpioInterface.html#a75e29a37bdcfdfdc2e42a7d556ef945b',1,'ace_segment::GpioInterface']]]
+  ['pinmode_117',['pinMode',['../classace__segment_1_1GpioInterface.html#a75e29a37bdcfdfdc2e42a7d556ef945b',1,'ace_segment::GpioInterface']]]
 ];