+++ docs/html/search/functions_c.js
@@ -1,10 +1,10 @@
 var searchData=
 [
-  ['scanningmodule_117',['ScanningModule',['../classace__segment_1_1ScanningModule.html#aa4e9ffaca62613fa76e885565716e5a4',1,'ace_segment::ScanningModule']]],
-  ['setbrightness_118',['setBrightness',['../classace__segment_1_1LedModule.html#ab6cb9fa9e4aa2af6a3ad134dba32ab59',1,'ace_segment::LedModule']]],
-  ['setbrightnessat_119',['setBrightnessAt',['../classace__segment_1_1ScanningModule.html#aea3b11435229958e33321c6e1d0e8553',1,'ace_segment::ScanningModule']]],
-  ['setbrightnessdirty_120',['setBrightnessDirty',['../classace__segment_1_1LedModule.html#abb8708d2e8733baa60d3e52ea178b20e',1,'ace_segment::LedModule']]],
-  ['setdigitdirty_121',['setDigitDirty',['../classace__segment_1_1LedModule.html#a04715054af6342e92499cb2811f535ad',1,'ace_segment::LedModule']]],
-  ['setdisplayon_122',['setDisplayOn',['../classace__segment_1_1Tm1637Module.html#ae86f7df1dc791d44feaad9253f6df184',1,'ace_segment::Tm1637Module::setDisplayOn()'],['../classace__segment_1_1Tm1638Module.html#a982da0b0660d2dfcd8e1fcfdb1e8fdb6',1,'ace_segment::Tm1638Module::setDisplayOn()']]],
-  ['setpatternat_123',['setPatternAt',['../classace__segment_1_1LedModule.html#a94f8a66e505096d137c78edc519b0386',1,'ace_segment::LedModule']]]
+  ['scanningmodule_121',['ScanningModule',['../classace__segment_1_1ScanningModule.html#aa4e9ffaca62613fa76e885565716e5a4',1,'ace_segment::ScanningModule']]],
+  ['setbrightness_122',['setBrightness',['../classace__segment_1_1LedModule.html#ab6cb9fa9e4aa2af6a3ad134dba32ab59',1,'ace_segment::LedModule']]],
+  ['setbrightnessat_123',['setBrightnessAt',['../classace__segment_1_1ScanningModule.html#aea3b11435229958e33321c6e1d0e8553',1,'ace_segment::ScanningModule']]],
+  ['setbrightnessdirty_124',['setBrightnessDirty',['../classace__segment_1_1LedModule.html#abb8708d2e8733baa60d3e52ea178b20e',1,'ace_segment::LedModule']]],
+  ['setdigitdirty_125',['setDigitDirty',['../classace__segment_1_1LedModule.html#a04715054af6342e92499cb2811f535ad',1,'ace_segment::LedModule']]],
+  ['setdisplayon_126',['setDisplayOn',['../classace__segment_1_1Tm1637Module.html#ae86f7df1dc791d44feaad9253f6df184',1,'ace_segment::Tm1637Module::setDisplayOn()'],['../classace__segment_1_1Tm1638AnodeModule.html#a49743e2eb9e819dc7ca97d069e26279a',1,'ace_segment::Tm1638AnodeModule::setDisplayOn()'],['../classace__segment_1_1Tm1638Module.html#a982da0b0660d2dfcd8e1fcfdb1e8fdb6',1,'ace_segment::Tm1638Module::setDisplayOn()']]],
+  ['setpatternat_127',['setPatternAt',['../classace__segment_1_1LedModule.html#a94f8a66e505096d137c78edc519b0386',1,'ace_segment::LedModule']]]
 ];