+++ docs/html/search/functions_d.js
@@ -1,5 +1,6 @@
 var searchData=
 [
-  ['tm1637module_124',['Tm1637Module',['../classace__segment_1_1Tm1637Module.html#a9cc401ceeed8292b26a93d5dcc08312e',1,'ace_segment::Tm1637Module']]],
-  ['tm1638module_125',['Tm1638Module',['../classace__segment_1_1Tm1638Module.html#a0fdb90f996a0f47838a76bc500550fd4',1,'ace_segment::Tm1638Module']]]
+  ['tm1637module_128',['Tm1637Module',['../classace__segment_1_1Tm1637Module.html#a9cc401ceeed8292b26a93d5dcc08312e',1,'ace_segment::Tm1637Module']]],
+  ['tm1638anodemodule_129',['Tm1638AnodeModule',['../classace__segment_1_1Tm1638AnodeModule.html#aa981626f917c6e7bacec2e2643042797',1,'ace_segment::Tm1638AnodeModule']]],
+  ['tm1638module_130',['Tm1638Module',['../classace__segment_1_1Tm1638Module.html#a0fdb90f996a0f47838a76bc500550fd4',1,'ace_segment::Tm1638Module']]]
 ];