+++ src/AceSegment.h
@@ -43,8 +43,8 @@ SOFTWARE.
 #endif
 
 // Version format: xxyyzz == "xx.yy.zz"
-#define ACE_SEGMENT_VERSION 1100
-#define ACE_SEGMENT_VERSION_STRING "0.11.0"
+#define ACE_SEGMENT_VERSION 1200
+#define ACE_SEGMENT_VERSION_STRING "0.12.0"
 
 #include "ace_segment/hw/ClockInterface.h"
 #include "ace_segment/hw/GpioInterface.h"
@@ -59,6 +59,7 @@ SOFTWARE.
 #include "ace_segment/hc595/Hc595Module.h"
 #include "ace_segment/tm1637/Tm1637Module.h"
 #include "ace_segment/tm1638/Tm1638Module.h"
+#include "ace_segment/tm1638/Tm1638AnodeModule.h"
 #include "ace_segment/max7219/Max7219Module.h"
 #include "ace_segment/ht16k33/Ht16k33Module.h"